Updated project structure and added new about me section
