+++ src/components/Home/index.module.css
@@ -1,0 +1,62 @@
+.intro-container {
+    display: grid;
+    grid-template-columns: 1fr 1fr;
+    grid-template-rows: 1fr;
+    height: 100%;
+}
+
+.div-one {
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    background-color: #2f4f4f;
+    color: #dae7e6;
+}
+
+
+.div-two {
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    background-color: #dae7e6;
+    color: #f2e9e4;
+}
+
+.text-div {
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    height: 50%;
+    padding: 1em 3.5em;
+}
+
+.main-heading {
+    font-size: 9.5rem;
+    font-family: "Prata", serif;
+    position: relative;
+}
+
+.sub-heading {
+    font-size: 2rem;
+    margin-top: 2%;
+}
+
+.imgDiv {
+    margin: 0 auto; 
+}
+
+.mePic {
+    border-radius: 50%;
+}
+
+.toggle-icon {
+    position: fixed;
+    top: 20px;
+    right: 20px;
+    width: 50px;
+    height: 50px;
+    border-radius: 50%;
+    z-index: 20;
+    cursor: pointer;
+}
