+++ src/pages/index.js
@@ -1,5 +1,6 @@
 import * as React from "react";
 import Layout from "../components/Layout";
+import Home from "../components/Home";
 import About from "../components/About";
 import Projects from "../components/Projects";
 import Contacts from "../components/Contact";
@@ -7,6 +8,7 @@ import Contacts from "../components/Contact";
 const IndexPage = () => {
   return (
       <Layout>
+        <Home />
         <About />
         <Projects />
         <Contacts />
